completed desktop view

+++ src/services/auth/index.js
@@ -14,11 +14,16 @@ const BASE_URL = import.meta.env.VITE_BASE_URL;
 export function login(formData, navigate, resetForm) {
   return async (dispatch) => {
     dispatch(setLoading(true));
-    Swal.showLoading();
+    Swal.fire({
+      title:"Processing..",
+      didOpen:()=>{
+        Swal.showLoading();
+      }
+    })
     const data = await Fetch(endpoints.LOGIN_API, "POST", formData);
 
     if (!data.success) {
-      Swal.hideLoading();
+      Swal.close();
       Swal.fire({
         position: "center",
         icon: "error",
@@ -29,7 +34,7 @@ export function login(formData, navigate, resetForm) {
       resetForm();
       return;
     } else {
-      Swal.hideLoading();
+      Swal.close();
       resetForm();
       dispatch(setToken(data.token));
       dispatch(setUser(data.user));
@@ -48,28 +53,42 @@ export function login(formData, navigate, resetForm) {
 export function logout(navigate) {
   return (dispatch) => {
     Swal.fire({
-      position: "center",
-      icon: "success",
-      title: "Logged out",
-      showConfirmButton: false,
-      timer: 1500,
+      title: "Are you sure?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Logout",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          position: "center",
+          icon: "success",
+          title: "Logged out",
+          showConfirmButton: false,
+          timer: 1500,
+        });
+        dispatch(setLoading(false));
+        dispatch(removeToken());
+        dispatch(setUser(null));
+        dispatch(setUrls(null));
+        navigate("/");
+      }
     });
-    dispatch(setLoading(false));
-    dispatch(removeToken());
-    dispatch(setUser(null));
-    dispatch(setUrls(null));
-    navigate("/");
   };
 }
 
 export function signup(formData, navigate, resetForm) {
   return async (dispatch) => {
     dispatch(setLoading(true));
-    Swal.showLoading();
+    Swal.fire({
+      title:"Creating account..",
+      didOpen:()=>{
+        Swal.showLoading();
+      }
+    })
 
     const data = await Fetch(endpoints.SINGUP_API, "POST", formData);
     if (data.success) {
-      Swal.hideLoading();
+      Swal.close();
       Swal.fire({
         position: "center",
         icon: "success",
@@ -81,7 +100,7 @@ export function signup(formData, navigate, resetForm) {
       resetForm();
       navigate("/profile");
     } else {
-      Swal.hideLoading();
+      Swal.close();
       Swal.fire({
         position: "center",
         icon: "error",
@@ -129,10 +148,15 @@ export function profile(token, navigate) {
 
 export function deleteProfile(token, navigate) {
   return async (dispatch) => {
-    Swal.showLoading();
+    Swal.fire({
+      title:"Deleting profile..",
+      didOpen:()=>{
+        Swal.showLoading();
+      }
+    })
     const data = await Fetch(endpoints.DELETE_PROFIE_API, "DELETE", {}, token);
     if (data.success) {
-      Swal.hideLoading();
+      Swal.close();
       Swal.fire({
         position: "center",
         icon: "success",
@@ -146,7 +170,7 @@ export function deleteProfile(token, navigate) {
       dispatch(setUrls(null));
       navigate("/");
     } else {
-      Swal.hideLoading();
+      Swal.close();
       Swal.fire({
         position: "center",
         icon: "success",
